+++ src/map/packets/merit_points_categories.cpp
@@ -77,7 +77,7 @@ CMeritPointsCategoriesPacket::CMeritPointsCategoriesPacket(CCharEntity* PChar)
 
         PChar->pushPacket(new CBasicPacket(*this));
     }
-    MeritPointsCategoriesPacket(PChar, 5 * MAX_MERITS_IN_PACKET);
+    MeritPointsCategoriesPacket(PChar, (uint8)(5 * MAX_MERITS_IN_PACKET));
 }
 
 /************************************************************************
